add rule to prevent accesses to upper packages
For some projects it makes sense to forbid all accesses to upper packages, as they point to implementation details being public in the API (given the root package of a component is your API by convention). This adds a rule to the library that forbids all such accesses within the project.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>