prepare test assertions for `JavaType`
Instead of `JavaClass(es)Assertion` we will need `JavaType(s)Assertion` in the future to be able to handle generic test cases. We can easily allow assertions on `JavaType` by simply pre-asserting `instanceof JavaClass` for the relevant cases specific to `JavaClass`. We could introduce a special sub-assertion `JavaClassAssertion` extending `JavaTypeAssertion`, where an assertion of the correct sub type is not necessary anymore, but to me the complexity seems to be fairly low to simply assert the correct sub type where necessary, so I kept it simple and only introduced `JavaType(s)Assertion`.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>